move combo code to files
