+++ combo/keycodes.h
@@ -1,0 +1,63 @@
+#if COMBO_KEYS_COUNT > 0
+  CMB_000,
+#endif
+#if COMBO_KEYS_COUNT > 1
+  CMB_001,
+#endif
+#if COMBO_KEYS_COUNT > 2
+  CMB_002,
+#endif
+#if COMBO_KEYS_COUNT > 3
+  CMB_003,
+#endif
+#if COMBO_KEYS_COUNT > 4
+  CMB_004,
+#endif
+#if COMBO_KEYS_COUNT > 5
+  CMB_005,
+#endif
+#if COMBO_KEYS_COUNT > 6
+  CMB_006,
+#endif
+#if COMBO_KEYS_COUNT > 7
+  CMB_007,
+#endif
+#if COMBO_KEYS_COUNT > 8
+  CMB_008,
+#endif
+#if COMBO_KEYS_COUNT > 9
+  CMB_009,
+#endif
+#if COMBO_KEYS_COUNT > 10
+  CMB_010,
+#endif
+#if COMBO_KEYS_COUNT > 11
+  CMB_011,
+#endif
+#if COMBO_KEYS_COUNT > 12
+  CMB_012,
+#endif
+#if COMBO_KEYS_COUNT > 13
+  CMB_013,
+#endif
+#if COMBO_KEYS_COUNT > 14
+  CMB_014,
+#endif
+#if COMBO_KEYS_COUNT > 15
+  CMB_015,
+#endif
+#if COMBO_KEYS_COUNT > 16
+  CMB_016,
+#endif
+#if COMBO_KEYS_COUNT > 17
+  CMB_017,
+#endif
+#if COMBO_KEYS_COUNT > 18
+  CMB_018,
+#endif
+#if COMBO_KEYS_COUNT > 19
+  CMB_019,
+#endif
+#if COMBO_KEYS_COUNT > 20
+  CMB_020,
+#endif
