+++ combo/process_record_user.c
@@ -1,0 +1,4 @@
+if (combo_enabled) {
+	if (combo_process(key, record))
+		return false;
+}
